Safari-609.1.11.1/LayoutTests/http/tests/security/dataURL/xss-DENIED-from-data-url-sub-frame-expected.txt

CONSOLE MESSAGE: line 1: SecurityError: Blocked a frame with origin "null" from accessing a cross-origin frame. Protocols, domains, and ports must match.
This tests that a data: URL loaded in an iframe doesn't have access to its parent's frame


PASS: Cross frame access from a data: URL was denied.

--------
Frame: 'aFrame'
--------
Inner iframe.
